// src/appearance/theme/types.tsx
export interface iTheme {
  backgroundColor: string;
  secondaryBackgroundColor: string;
  foregroundColor: string;
  secondaryForegroundColor: string;
  buttonBackgroundColor: string;
  buttonSecondaryBackgroundColor: string;
  buttonThirdBackgroundColor: string;
  borderColor: string;
  iconColor: string;
  buttonBorderColor: string;
  textColor: string;
  buttonTextColor: string;
  secondaryTextColor: string;
  rippleColor: string;
  tabBorderColor: string;
}